+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 * Use Appsignal.monitor_transaction to instrument and log errors for
   custom actions
 * Add option to ignore a specific action
+* Use a Mutex in agent instead of Thread.exclusive
 
 # 0.9.3
 * Log if appsignal is not active for an environment